+++ NumLib/ODESolver/TimeDiscretizedODESystem.h
@@ -12,32 +12,10 @@
 
 #include <memory>
 
-#include "MathLib/LinAlg/ApplyKnownSolution.h"
-#include "MathLib/LinAlg/UnifiedMatrixSetters.h"
-#include "NumLib/IndexValueVector.h"
-
 #include "ODESystem.h"
 #include "NonlinearSystem.h"
 #include "TimeDiscretization.h"
 #include "MatrixTranslator.h"
-
-namespace detail
-{
-//! Applies known solutions to the solution vector \c x.
-template<typename Solutions, typename Vector>
-void applyKnownSolutions(std::vector<Solutions> const*const known_solutions,
-                         Vector& x)
-{
-    if (!known_solutions) return;
-
-    for (auto const& bc : *known_solutions) {
-        for (std::size_t i=0; i<bc.ids.size(); ++i) {
-            // TODO that might have bad performance for some Vector types, e.g., PETSc.
-            MathLib::setVector(x, bc.ids[i], bc.values[i]);
-        }
-    }
-}
-}
 
 namespace NumLib
 {
@@ -109,112 +87,22 @@ public:
      * \param time_discretization the time discretization to be used.
      */
     explicit
-    TimeDiscretizedODESystem(ODE& ode, TimeDisc& time_discretization)
-        : _ode(ode)
-        , _time_disc(time_discretization)
-        , _mat_trans(createMatrixTranslator<ODETag>(time_discretization))
-    {
-        _Jac  = &MathLib::GlobalMatrixProvider::provider.getMatrix(
-                    _ode.getMatrixSpecifications(), _Jac_id);
-        _M    = &MathLib::GlobalMatrixProvider::provider.getMatrix(
-                    _ode.getMatrixSpecifications(), _M_id);
-        _K    = &MathLib::GlobalMatrixProvider::provider.getMatrix(
-                    _ode.getMatrixSpecifications(), _K_id);
-        _b    = &MathLib::GlobalVectorProvider::provider.getVector(
-                    _ode.getMatrixSpecifications(), _b_id);
-    }
+    TimeDiscretizedODESystem(ODE& ode, TimeDisc& time_discretization);
 
-    ~TimeDiscretizedODESystem()
-    {
-        MathLib::GlobalMatrixProvider::provider.releaseMatrix(*_Jac);
-        MathLib::GlobalMatrixProvider::provider.releaseMatrix(*_M);
-        MathLib::GlobalMatrixProvider::provider.releaseMatrix(*_K);
-        MathLib::GlobalVectorProvider::provider.releaseVector(*_b);
-    }
+    ~TimeDiscretizedODESystem();
 
-    void assembleResidualNewton(const GlobalVector &x_new_timestep) override
-    {
-        namespace BLAS = MathLib::BLAS;
+    void assembleResidualNewton(const GlobalVector &x_new_timestep) override;
 
-        auto const  t      = _time_disc.getCurrentTime();
-        auto const& x_curr = _time_disc.getCurrentX(x_new_timestep);
+    void assembleJacobian(const GlobalVector &x_new_timestep) override;
 
-        _M->setZero();
-        _K->setZero();
-        _b->setZero();
+    void getResidual(GlobalVector const& x_new_timestep, GlobalVector& res) const override;
 
-        _ode.assemble(t, x_curr, *_M, *_K, *_b);
+    void getJacobian(GlobalMatrix& Jac) const override;
 
-        BLAS::finalizeAssembly(*_M);
-        BLAS::finalizeAssembly(*_K);
-        BLAS::finalizeAssembly(*_b);
-    }
-
-    void assembleJacobian(const GlobalVector &x_new_timestep) override
-    {
-        namespace BLAS = MathLib::BLAS;
-
-        auto const  t        = _time_disc.getCurrentTime();
-        auto const& x_curr   = _time_disc.getCurrentX(x_new_timestep);
-        auto const  dxdot_dx = _time_disc.getNewXWeight();
-        auto const  dx_dx    = _time_disc.getDxDx();
-
-        auto& xdot = MathLib::GlobalVectorProvider::provider.getVector(_xdot_id);
-        _time_disc.getXdot(x_new_timestep, xdot);
-
-        _Jac->setZero();
-
-        _ode.assembleJacobian(t, x_curr, xdot,
-                              dxdot_dx, *_M, dx_dx, *_K,
-                              *_Jac);
-
-        MathLib::BLAS::finalizeAssembly(*_Jac);
-
-        MathLib::GlobalVectorProvider::provider.releaseVector(xdot);
-    }
-
-    void getResidual(GlobalVector const& x_new_timestep, GlobalVector& res) const override
-    {
-        // TODO Maybe the duplicate calculation of xdot here and in assembleJacobian
-        //      can be optimuized. However, that would make the interface a bit more
-        //      fragile.
-        auto& xdot = MathLib::GlobalVectorProvider::provider.getVector(_xdot_id);
-        _time_disc.getXdot(x_new_timestep, xdot);
-
-        _mat_trans->computeResidual(*_M, *_K, *_b, x_new_timestep, xdot, res);
-
-        MathLib::GlobalVectorProvider::provider.releaseVector(xdot);
-    }
-
-    void getJacobian(GlobalMatrix& Jac) const override
-    {
-        _mat_trans->computeJacobian(*_Jac, Jac);
-    }
-
-    void applyKnownSolutions(GlobalVector& x) const override
-    {
-        ::detail::applyKnownSolutions(
-                    _ode.getKnownSolutions(_time_disc.getCurrentTime()), x);
-    }
+    void applyKnownSolutions(GlobalVector& x) const override;
 
     void applyKnownSolutionsNewton(GlobalMatrix& Jac, GlobalVector& res,
-                                    GlobalVector& minus_delta_x) override
-    {
-        auto const* known_solutions =
-            _ode.getKnownSolutions(_time_disc.getCurrentTime());
-
-        if (known_solutions) {
-            std::vector<double> values;
-
-            for (auto const& bc : *known_solutions) {
-                // TODO this is the quick and dirty and bad performance solution.
-                values.resize(bc.values.size(), 0.0);
-
-                // TODO maybe it would be faster to apply all at once
-                MathLib::applyKnownSolution(Jac, res, minus_delta_x, bc.ids, values);
-            }
-        }
-    }
+                                    GlobalVector& minus_delta_x) override;
 
     bool isLinear() const override
     {
@@ -296,44 +184,11 @@ public:
      * \param time_discretization the time discretization to be used.
      */
     explicit
-    TimeDiscretizedODESystem(ODE& ode, TimeDisc& time_discretization)
-        : _ode(ode)
-        , _time_disc(time_discretization)
-        , _mat_trans(createMatrixTranslator<ODETag>(
-                         time_discretization))
-    {
-        _M = &MathLib::GlobalMatrixProvider::provider.getMatrix(
-                    ode.getMatrixSpecifications(), _M_id);
-        _K = &MathLib::GlobalMatrixProvider::provider.getMatrix(
-                    ode.getMatrixSpecifications(), _K_id);
-        _b = &MathLib::GlobalVectorProvider::provider.getVector(
-                    ode.getMatrixSpecifications(), _b_id);
-    }
+    TimeDiscretizedODESystem(ODE& ode, TimeDisc& time_discretization);
 
-    ~TimeDiscretizedODESystem()
-    {
-        MathLib::GlobalMatrixProvider::provider.releaseMatrix(*_M);
-        MathLib::GlobalMatrixProvider::provider.releaseMatrix(*_K);
-        MathLib::GlobalVectorProvider::provider.releaseVector(*_b);
-    }
+    ~TimeDiscretizedODESystem();
 
-    void assembleMatricesPicard(const GlobalVector &x_new_timestep) override
-    {
-        namespace BLAS = MathLib::BLAS;
-
-        auto const  t      = _time_disc.getCurrentTime();
-        auto const& x_curr = _time_disc.getCurrentX(x_new_timestep);
-
-        _M->setZero();
-        _K->setZero();
-        _b->setZero();
-
-        _ode.assemble(t, x_curr, *_M, *_K, *_b);
-
-        BLAS::finalizeAssembly(*_M);
-        BLAS::finalizeAssembly(*_K);
-        BLAS::finalizeAssembly(*_b);
-    }
+    void assembleMatricesPicard(const GlobalVector &x_new_timestep) override;
 
     void getA(GlobalMatrix& A) const override
     {
@@ -345,28 +200,9 @@ public:
         _mat_trans->computeRhs(*_M, *_K, *_b, rhs);
     }
 
-    void applyKnownSolutions(GlobalVector& x) const override
-    {
-        ::detail::applyKnownSolutions(
-                    _ode.getKnownSolutions(_time_disc.getCurrentTime()), x);
-    }
+    void applyKnownSolutions(GlobalVector& x) const override;
 
-    void applyKnownSolutionsPicard(GlobalMatrix& A, GlobalVector& rhs, GlobalVector& x) override
-    {
-        auto const* known_solutions =
-            _ode.getKnownSolutions(_time_disc.getCurrentTime());
-
-        if (known_solutions) {
-            using IndexType = typename MathLib::MatrixVectorTraits<GlobalMatrix>::Index;
-            std::vector<IndexType> ids;
-            std::vector<double> values;
-            for (auto const& bc : *known_solutions) {
-                std::copy(bc.ids.cbegin(), bc.ids.cend(), std::back_inserter(ids));
-                std::copy(bc.values.cbegin(), bc.values.cend(), std::back_inserter(values));
-            }
-            MathLib::applyKnownSolution(A, rhs, x, ids, values);
-        }
-    }
+    void applyKnownSolutionsPicard(GlobalMatrix& A, GlobalVector& rhs, GlobalVector& x) override;
 
     bool isLinear() const override
     {
